refactor CA generation
