+++ src/model/app_data.rs
@@ -1,6 +1,7 @@
 use crate::config::Config;
 use crate::error::Error;
 use crate::model::audience::AudiencesStore;
+use crate::model::ca::CA;
 use crate::model::jwks::JwksStore;
 
 use super::authorizations::Authorizations;
@@ -10,14 +11,17 @@ pub struct AppData {
     audiences_store: AudiencesStore,
     jwks_store: JwksStore,
     authorizations: Authorizations,
+    ca: CA,
 }
 
 impl AppData {
     pub fn new(config: Config) -> Result<Self, Error> {
+        let ca = CA::new()?;
         Ok(Self {
             audiences_store: AudiencesStore::new(config.audience()),
-            jwks_store: JwksStore::new()?,
+            jwks_store: JwksStore::new(&ca)?,
             authorizations: Authorizations::default(),
+            ca,
             config,
         })
     }
@@ -36,5 +40,9 @@ impl AppData {
 
     pub fn authorizations(&self) -> &Authorizations {
         &self.authorizations
+    }
+
+    pub fn ca(&self) -> &CA {
+        &self.ca
     }
 }
