+++ src/model/jwks.rs
@@ -2,31 +2,24 @@ use std::str::FromStr;
 use std::sync::{RwLock, RwLockWriteGuard};
 
 use jsonwebtoken::{Algorithm, DecodingKey, EncodingKey, Header, Validation};
-use openssl::asn1::Asn1Time;
-use openssl::bn::{BigNum, MsbOption};
-use openssl::error::ErrorStack;
-use openssl::hash::MessageDigest;
 use openssl::pkey::{PKey, Private};
 use openssl::rsa::Rsa;
-use openssl::x509::extension::{
-    AuthorityKeyIdentifier, BasicConstraints, KeyUsage, SubjectAlternativeName, SubjectKeyIdentifier,
-};
-use openssl::x509::{X509NameBuilder, X509Req, X509ReqBuilder, X509};
 use rand::seq::SliceRandom;
 use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
 use crate::error::Error;
+use crate::model::ca::CA;
 
 pub struct JwksStore {
     cache: RwLock<Jwks>,
 }
 
 impl JwksStore {
-    pub fn new() -> Result<Self, Error> {
+    pub fn new(ca: &CA) -> Result<Self, Error> {
         Ok(Self {
-            cache: RwLock::new(Jwks::new()?),
+            cache: RwLock::new(Jwks::new(ca)?),
         })
     }
 
@@ -38,15 +31,15 @@ impl JwksStore {
         self.cache.read().unwrap_or_else(|p| p.into_inner()).random_jwk()
     }
 
-    pub fn rotate_keys(&self) -> Result<(), Error> {
+    pub fn rotate_keys(&self, ca: &CA) -> Result<(), Error> {
         let mut jwks: RwLockWriteGuard<Jwks> = self.cache.write().unwrap_or_else(|p| p.into_inner());
-        *jwks = jwks.rotate_keys()?;
+        *jwks = jwks.rotate_keys(ca)?;
         Ok(())
     }
 
-    pub fn revoke_keys(&self) -> Result<(), Error> {
+    pub fn revoke_keys(&self, ca: &CA) -> Result<(), Error> {
         let mut jwks: RwLockWriteGuard<Jwks> = self.cache.write().unwrap_or_else(|p| p.into_inner());
-        *jwks = jwks.revoke_keys()?;
+        *jwks = jwks.revoke_keys(ca)?;
         Ok(())
     }
 }
@@ -57,11 +50,11 @@ pub struct Jwks {
 }
 
 impl Jwks {
-    pub fn new() -> Result<Self, Error> {
+    pub fn new(ca: &CA) -> Result<Self, Error> {
         Ok(Self {
             keys: (1..=3)
                 .into_iter()
-                .map(|_| Jwk::new(None))
+                .map(|_| Jwk::new(None, ca))
                 .collect::<Result<Vec<Jwk>, Error>>()?,
         })
     }
@@ -94,15 +87,15 @@ impl Jwks {
         }
     }
 
-    pub fn rotate_keys(&self) -> Result<Self, Error> {
+    pub fn rotate_keys(&self, ca: &CA) -> Result<Self, Error> {
         let mut keys: Vec<Jwk> = self.keys.clone();
-        keys.insert(0, Jwk::new(None)?);
+        keys.insert(0, Jwk::new(None, ca)?);
         keys.pop();
         Ok(Self { keys })
     }
 
-    pub fn revoke_keys(&self) -> Result<Self, Error> {
-        Jwks::new()
+    pub fn revoke_keys(&self, ca: &CA) -> Result<Self, Error> {
+        Jwks::new(ca)
     }
 }
 
@@ -120,7 +113,7 @@ pub struct Jwk {
 }
 
 impl Jwk {
-    pub fn new(rsa: Option<Rsa<Private>>) -> Result<Jwk, Error> {
+    pub fn new(rsa: Option<Rsa<Private>>, ca: &CA) -> Result<Jwk, Error> {
         let rsa: Rsa<Private> = match rsa {
             Some(rsa) => rsa,
             None => Rsa::generate(2048)?,
@@ -130,7 +123,7 @@ impl Jwk {
         let exponent: Vec<u8> = rsa.e().to_vec();
         let key_pair: PKey<Private> = PKey::from_rsa(rsa)?;
 
-        let (x509, key_pair) = make_x509_cert(key_pair)?;
+        let x509 = ca.sign(&key_pair)?;
         let x509pem = x509.to_pem()?;
         let x509cert = String::from_utf8(x509pem)?
             .replace('\n', "")
@@ -181,122 +174,21 @@ impl Jwk {
     }
 }
 
-fn make_x509_request(key_pair: &PKey<Private>) -> Result<X509Req, ErrorStack> {
-    let mut req_builder = X509ReqBuilder::new()?;
-    req_builder.set_pubkey(key_pair)?;
-
-    let mut x509_name = X509NameBuilder::new()?;
-    x509_name.append_entry_by_text("C", "IT")?;
-    x509_name.append_entry_by_text("O", "Prima")?;
-    x509_name.append_entry_by_text("CN", "prima.localauth0.com")?;
-    let x509_name = x509_name.build();
-    req_builder.set_subject_name(&x509_name)?;
-
-    req_builder.sign(key_pair, MessageDigest::sha256())?;
-    let req = req_builder.build();
-    Ok(req)
-}
-
-fn make_ca_cert() -> Result<(X509, PKey<Private>), ErrorStack> {
-    let rsa = Rsa::generate(2048)?;
-    let key_pair = PKey::from_rsa(rsa)?;
-
-    let mut x509_name = X509NameBuilder::new()?;
-    x509_name.append_entry_by_text("C", "IT")?;
-    x509_name.append_entry_by_text("O", "Prima CA")?;
-    x509_name.append_entry_by_text("CN", "Prima CA")?;
-    let x509_name = x509_name.build();
-
-    let mut cert_builder = X509::builder()?;
-    cert_builder.set_version(2)?;
-    let serial_number = {
-        let mut serial = BigNum::new()?;
-        serial.rand(159, MsbOption::MAYBE_ZERO, false)?;
-        serial.to_asn1_integer()?
-    };
-    cert_builder.set_serial_number(&serial_number)?;
-    cert_builder.set_subject_name(&x509_name)?;
-    cert_builder.set_issuer_name(&x509_name)?;
-    cert_builder.set_pubkey(&key_pair)?;
-    let not_before = Asn1Time::days_from_now(0)?;
-    cert_builder.set_not_before(&not_before)?;
-    let not_after = Asn1Time::days_from_now(365)?;
-    cert_builder.set_not_after(&not_after)?;
-
-    cert_builder.append_extension(BasicConstraints::new().critical().ca().build()?)?;
-    cert_builder.append_extension(KeyUsage::new().critical().key_cert_sign().crl_sign().build()?)?;
-
-    let subject_key_identifier = SubjectKeyIdentifier::new().build(&cert_builder.x509v3_context(None, None))?;
-    cert_builder.append_extension(subject_key_identifier)?;
-
-    cert_builder.sign(&key_pair, MessageDigest::sha256())?;
-    let cert = cert_builder.build();
-
-    Ok((cert, key_pair))
-}
-
-fn make_x509_cert(key_pair: PKey<Private>) -> Result<(X509, PKey<Private>), ErrorStack> {
-    let (ca_cert, ca_key_pair) = make_ca_cert()?;
-    let req = make_x509_request(&key_pair)?;
-
-    let mut cert_builder = X509::builder()?;
-    cert_builder.set_version(2)?;
-    let serial_number = {
-        let mut serial = BigNum::new()?;
-        serial.rand(159, MsbOption::MAYBE_ZERO, false)?;
-        serial.to_asn1_integer()?
-    };
-    cert_builder.set_serial_number(&serial_number)?;
-    cert_builder.set_subject_name(req.subject_name())?;
-    cert_builder.set_issuer_name(ca_cert.subject_name())?;
-    cert_builder.set_pubkey(&key_pair)?;
-    let not_before = Asn1Time::days_from_now(0)?;
-    cert_builder.set_not_before(&not_before)?;
-    let not_after = Asn1Time::days_from_now(365)?;
-    cert_builder.set_not_after(&not_after)?;
-
-    cert_builder.append_extension(BasicConstraints::new().build()?)?;
-
-    cert_builder.append_extension(
-        KeyUsage::new()
-            .critical()
-            .non_repudiation()
-            .digital_signature()
-            .key_encipherment()
-            .build()?,
-    )?;
-
-    let subject_key_identifier =
-        SubjectKeyIdentifier::new().build(&cert_builder.x509v3_context(Some(&ca_cert), None))?;
-    cert_builder.append_extension(subject_key_identifier)?;
-
-    let auth_key_identifier = AuthorityKeyIdentifier::new()
-        .keyid(false)
-        .issuer(false)
-        .build(&cert_builder.x509v3_context(Some(&ca_cert), None))?;
-    cert_builder.append_extension(auth_key_identifier)?;
-
-    let subject_alt_name = SubjectAlternativeName::new()
-        .dns("auth0-proxy")
-        .dns("auth0")
-        .build(&cert_builder.x509v3_context(Some(&ca_cert), None))?;
-    cert_builder.append_extension(subject_alt_name)?;
-
-    cert_builder.sign(&ca_key_pair, MessageDigest::sha256())?;
-    let cert = cert_builder.build();
-
-    Ok((cert, key_pair))
-}
-
 #[cfg(test)]
 mod tests {
     use crate::error::Error;
+    use crate::model::ca::CA;
     use crate::model::jwks::JwksStore;
     use crate::model::{Claims, GrantType, Jwk, Jwks};
 
     #[test]
     fn its_possible_to_generate_jwks_and_parse_claims_using_given_jwks_test() {
-        let jwk_store: JwksStore = JwksStore::new().unwrap();
+        let ca = match CA::new() {
+            Ok(ca) => ca,
+            _ => panic!("Cannot initialize CA"),
+        };
+
+        let jwk_store: JwksStore = JwksStore::new(&ca).unwrap();
         let audience: &str = "audience";
         let permission: &str = "permission";
         let issuer: &str = "issuer";
