+++ src/model/mod.rs
@@ -8,6 +8,7 @@ pub use user_info::*;
 mod app_data;
 mod audience;
 mod authorizations;
+mod ca;
 mod claims;
 pub mod defaults;
 mod jwks;
